feat: align test layout with palo alto pattern via zscaler skeleton
- Adopt the zscaler/terraform-modules-zscaler-tests-skeleton@v0.0.3 helper
  library so per-folder main_test.go files stay tiny.
- Add TestValidate stubs to every modules/<x>/ (6 files) so the shared CI
  workflow's `make modules/<x> ACTION=Validate` matrix has a Go test to run
  in the same directory as the .tf files (matching PaO convention).
- Add full lifecycle (Validate / Plan / Apply / Idempotence) tests to every
  examples/<x>/ (4 files), driven by the skeleton's helpers; CI passes
  TF_VAR_name_prefix per run for collision-free naming.
- Replace the dual-style Makefile with PaO's single pattern target. The
  shared CI is now the sole entry point.
- Remove the orphaned test/ directory (logic now lives next to the
  modules/examples it exercises).

Verified locally: go mod tidy, go build, go vet, pre-commit terraform_docs
all clean; TestValidate passes against modules/terraform-zpa-app-connector-group.
