+++ examples/ac/main_test.go
@@ -1,0 +1,49 @@
+package ac
+
+import (
+	"log"
+	"testing"
+
+	"github.com/gruntwork-io/terratest/modules/logger"
+	"github.com/gruntwork-io/terratest/modules/terraform"
+	"github.com/zscaler/terraform-modules-zscaler-tests-skeleton/pkg/testskeleton"
+)
+
+func CreateTerraformOptions(t *testing.T) *terraform.Options {
+	varsInfo, err := testskeleton.GenerateTerraformVarsInfo("gcp")
+	if err != nil {
+		log.Fatalf("Error generating terraform vars info: %v", err)
+	}
+
+	return terraform.WithDefaultRetryableErrors(t, &terraform.Options{
+		TerraformDir: ".",
+		VarFiles:     []string{"terraform.tfvars"},
+		Vars: map[string]interface{}{
+			"name_prefix": varsInfo.NamePrefix,
+			"project":     varsInfo.GoogleProjectId,
+		},
+		Logger:               logger.Default,
+		Lock:                 true,
+		Upgrade:              true,
+		SetVarsAfterVarFiles: true,
+	})
+}
+
+func TestValidate(t *testing.T) {
+	testskeleton.ValidateCode(t, nil)
+}
+
+func TestPlan(t *testing.T) {
+	testskeleton.PlanInfraCheckErrors(t, CreateTerraformOptions(t),
+		[]testskeleton.AssertExpression{}, "No errors are expected")
+}
+
+func TestApply(t *testing.T) {
+	testskeleton.DeployInfraCheckOutputs(t, CreateTerraformOptions(t),
+		[]testskeleton.AssertExpression{})
+}
+
+func TestIdempotence(t *testing.T) {
+	testskeleton.DeployInfraCheckOutputsVerifyChanges(t, CreateTerraformOptions(t),
+		[]testskeleton.AssertExpression{})
+}
