+++ modules/terraform-zpa-app-connector-group/main_test.go
@@ -1,0 +1,11 @@
+package terraform_zpa_app_connector_group
+
+import (
+	"testing"
+
+	"github.com/zscaler/terraform-modules-zscaler-tests-skeleton/pkg/testskeleton"
+)
+
+func TestValidate(t *testing.T) {
+	testskeleton.ValidateCode(t, nil)
+}
