+++ modules/terraform-zpa-provisioning-key/main_test.go
@@ -1,0 +1,11 @@
+package terraform_zpa_provisioning_key
+
+import (
+	"testing"
+
+	"github.com/zscaler/terraform-modules-zscaler-tests-skeleton/pkg/testskeleton"
+)
+
+func TestValidate(t *testing.T) {
+	testskeleton.ValidateCode(t, nil)
+}
