+++ modules/terraform-zsac-acvm-gcp/main_test.go
@@ -1,0 +1,11 @@
+package terraform_zsac_acvm_gcp
+
+import (
+	"testing"
+
+	"github.com/zscaler/terraform-modules-zscaler-tests-skeleton/pkg/testskeleton"
+)
+
+func TestValidate(t *testing.T) {
+	testskeleton.ValidateCode(t, nil)
+}
