+++ modules/terraform-zsac-asg-gcp/main_test.go
@@ -1,0 +1,11 @@
+package terraform_zsac_asg_gcp
+
+import (
+	"testing"
+
+	"github.com/zscaler/terraform-modules-zscaler-tests-skeleton/pkg/testskeleton"
+)
+
+func TestValidate(t *testing.T) {
+	testskeleton.ValidateCode(t, nil)
+}
