+++ modules/terraform-zsac-bastion-gcp/main_test.go
@@ -1,0 +1,11 @@
+package terraform_zsac_bastion_gcp
+
+import (
+	"testing"
+
+	"github.com/zscaler/terraform-modules-zscaler-tests-skeleton/pkg/testskeleton"
+)
+
+func TestValidate(t *testing.T) {
+	testskeleton.ValidateCode(t, nil)
+}
