--- test/terraform-zpa-app-connector-group/test.tfvars
@@ -1,7 +1,0 @@
-# Test inputs for terraform-zpa-app-connector-group.
-# Provider credentials come from environment variables (see Makefile help).
-app_connector_group_enabled      = true
-app_connector_group_latitude     = "37.33874"
-app_connector_group_longitude    = "-121.8852525"
-app_connector_group_location     = "San Jose, CA, USA"
-app_connector_group_country_code = "US"
