--- test/terraform-zpa-provisioning-key/.terraform.lock.hcl
@@ -1,44 +1,0 @@
-# This file is maintained automatically by "terraform init".
-# Manual edits may be lost in future updates.
-
-provider "registry.terraform.io/hashicorp/random" {
-  version     = "3.8.1"
-  constraints = "~> 3.8.0"
-  hashes = [
-    "h1:u8AKlWVDTH5r9YLSeswoVEjiY72Rt4/ch7U+61ZDkiQ=",
-    "zh:08dd03b918c7b55713026037c5400c48af5b9f468f483463321bd18e17b907b4",
-    "zh:0eee654a5542dc1d41920bbf2419032d6f0d5625b03bd81339e5b33394a3e0ae",
-    "zh:229665ddf060aa0ed315597908483eee5b818a17d09b6417a0f52fd9405c4f57",
-    "zh:2469d2e48f28076254a2a3fc327f184914566d9e40c5780b8d96ebf7205f8bc0",
-    "zh:37d7eb334d9561f335e748280f5535a384a88675af9a9eac439d4cfd663bcb66",
-    "zh:741101426a2f2c52dee37122f0f4a2f2d6af6d852cb1db634480a86398fa3511",
-    "zh:78d5eefdd9e494defcb3c68d282b8f96630502cac21d1ea161f53cfe9bb483b3",
-    "zh:a902473f08ef8df62cfe6116bd6c157070a93f66622384300de235a533e9d4a9",
-    "zh:b85c511a23e57a2147355932b3b6dce2a11e856b941165793a0c3d7578d94d05",
-    "zh:c5172226d18eaac95b1daac80172287b69d4ce32750c82ad77fa0768be4ea4b8",
-    "zh:dab4434dba34aad569b0bc243c2d3f3ff86dd7740def373f2a49816bd2ff819b",
-    "zh:f49fd62aa8c5525a5c17abd51e27ca5e213881d58882fd42fec4a545b53c9699",
-  ]
-}
-
-provider "registry.terraform.io/zscaler/zpa" {
-  version     = "4.4.3"
-  constraints = "~> 4.4.0"
-  hashes = [
-    "h1:4A1/+x5y8zk3PAZ8xogLfX+JvDebdd4sI5vGfiJlZIw=",
-    "zh:3d6994923cb4ff9231de683076abcabfa1e1ec13e7a4d872226da4f641800f1d",
-    "zh:3dc9aca5b457898dc2e74d7ad16d41f6a69d2d7a4a1391e22aaa4fad050948a5",
-    "zh:73d7d6ab766110f76a6ff1fd36dbd32bcdcfc45460f904f8e015fdbd01a3f0d9",
-    "zh:87385c447c9a9eaf86617b9b359bfff46b3f34369dc1888b6d5d42ebed4cfbba",
-    "zh:929d8a01293c6226329a64ba2580ae2ac74e94c033997d86687ab14be1b48b4c",
-    "zh:9c21a5db9de41ecde9fcd71249eeaca892c66937b066bb5ef9aa720494390176",
-    "zh:b275710fa244503f114a69e8c637e1f2c5ad15ab7ba4bba73bd841f7ba45e686",
-    "zh:cb3f60f55122ca5d926caa5468836a8db5b3a3b525ef6d74ab661e9a460e4e57",
-    "zh:d23f3797c148929cc974b549c474b88aecbbb07cbe5e68f388a093d92e6223a3",
-    "zh:d7993eca61f2145e6246b5477a40f69d2454bd0ac33147acf7db34b16dd23508",
-    "zh:dff991fbf158a6495fec6ccb5ce3b5d13cc6e06e6f5a00aafb6e7e233f8ba292",
-    "zh:ef01b400b1cc48a0ee65bce5a883e5a0e9484d25f9ed173b499aa65e69bee08f",
-    "zh:efa425c56de1817a9d9dd17b29bfad2550449884926ef2ddbbd2cd269a63038b",
-    "zh:fbb6ba1ef0a7909cdc96041f7de4115572780277c7af5c309e5ddf7f15e5f804",
-  ]
-}
