--- test/terraform-zpa-provisioning-key/outputs.tf
@@ -1,26 +1,0 @@
-################################################################################
-# Passthrough outputs
-################################################################################
-output "app_connector_group_id" {
-  description = "ID of the App Connector Group the key is bound to"
-  value       = module.app_connector_group.app_connector_group_id
-}
-
-output "provisioning_key" {
-  description = "Provisioning key value generated by the module (sensitive)"
-  value       = module.provisioning_key.provisioning_key
-  sensitive   = true
-}
-
-################################################################################
-# Validation outputs
-################################################################################
-output "app_connector_group_id_valid" {
-  description = "Validation that the upstream App Connector Group ID is non-empty"
-  value       = length(module.app_connector_group.app_connector_group_id) > 0 ? "true" : "false"
-}
-
-output "provisioning_key_valid" {
-  description = "Validation that the provisioning key returned by the module is non-empty"
-  value       = length(module.provisioning_key.provisioning_key) > 0 ? "true" : "false"
-}
