--- test/terraform-zpa-provisioning-key/test.tfvars
@@ -1,5 +1,0 @@
-# Test inputs for terraform-zpa-provisioning-key.
-# Provider credentials come from environment variables (see Makefile help).
-provisioning_key_enabled          = true
-provisioning_key_association_type = "CONNECTOR_GRP"
-provisioning_key_max_usage        = 10
