--- test/terraform-zsac-acvm-gcp/test.tfvars
@@ -1,5 +1,0 @@
-# Test inputs for terraform-zsac-acvm-gcp.
-# project + region come from TF_VAR_project / TF_VAR_region (set by Makefile).
-az_count           = 1
-ac_count           = 1
-acvm_instance_type = "n2-standard-4"
